principal/requirements.txt

asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.9
django-autoslug==1.9.8
django-environ==0.8.1
django-heroku==0.3.1
django-on-heroku==1.1.2
django-autoslug==1.9.8
django-debug-toolbar==3.1.1
django-model-utils==4.1.0

mercadopago==2.0.8

environ==1.0
jsonfield==3.1.0

py==1.10.0
requests==2.25.1
selenium==4.1.0
urllib3==1.26.7

ipython==7.19.0

requests==2.25.1
jsonfield
gunicorn==20.1.0

pytest==6.1.2
python-decouple==3.5
pillow==8.0.1
psycopg2==2.8.6
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-mock==3.5.1